requirements.txt
numpy==1.16.1
pandas==0.24.1
python-dateutil==2.8.0
python-igraph==0.7.1.post6
pytz==2018.9
scikit-learn==0.20.2
scipy==1.2.1
six==1.12.0
